feat: initial build except add/edit errors

+++ controllers/character.js
@@ -1,5 +1,6 @@
 const { body, validationResult } = require('express-validator');
 const Character = require('../models/character');
+const Figure = require('../models/figure');
 
 exports.characterList = async (req, res, next) => {
 	try {
@@ -30,9 +31,11 @@ exports.characterDetail = async (req, res, next) => {
 };
 
 exports.characterCreateView = async (req, res, next) => {
-	console.log(req.params.id);
 	try {
-		res.render('character_create', { title: 'Create a New Character' });
+		res.render('character_form', { 
+			title: 'Create a New Character', 
+			actionType: 'Create',
+		});
 	} catch (err) {
 		next(err)
 	}
@@ -55,8 +58,8 @@ exports.characterCreate = [
 		const errors = validationResult(req);
 
 		if (!errors.isEmpty()) {
-			res.render('character_create', {
-				title: 'Create Character',
+			res.render('character_form', {
+				title: 'Create a New Character',
 				character: req.body,
 				errors: errors.array(),
 			});
@@ -78,8 +81,7 @@ exports.characterCreate = [
 	},
 ];
 
-exports.characterUpdateView = async (req, res
-	, next) => {
+exports.characterEditView = async (req, res, next) => {
 	try {
 		const character = await Character.findById(req.params.id);
 		if (!character) {
@@ -87,13 +89,18 @@ exports.characterUpdateView = async (req, res
 			err.status = 404;
 			return next(err);
 		}
-		res.render('character_update', { title: 'Update Character', character });
+
+		res.render('character_form', {
+			title: `Edit Character: ${character.name}`,
+			actionType: 'Edit',
+			character
+		});
 	} catch (err) {
 		next(err);
 	}
 };
 
-exports.characterUpdate = [
+exports.characterEdit = [
 	body('name', 'Name is required')
 		.trim()
 		.isLength({ min: 1 })
@@ -110,8 +117,9 @@ exports.characterUpdate = [
 		const errors = validationResult(req);
 
 		if (!errors.isEmpty()) {
-			res.render('character_update', {
-				title: 'Update Character',
+			res.render('character_form', {
+				title: `Edit Character: ${req.body.name}`,
+				action: 'Edit',
 				character: req.body,
 				errors: errors.array(),
 			});
@@ -141,7 +149,11 @@ exports.characterDeletionView = async (req, res, next) => {
 			err.status = 404;
 			return next(err);
 		}
-		res.render('character_delete', { title: 'Delete Character', character });
+		res.render('character_delete', { 
+			title: 'Delete Character', 
+			character,
+			figures: undefined 
+		});
 	} catch (err) {
 		next(err);
 	}
@@ -149,8 +161,20 @@ exports.characterDeletionView = async (req, res, next) => {
 
 exports.characterDelete = async (req, res, next) => {
 	try {
-		await Character.findByIdAndRemove(req.params.id);
-		res.status(204).redirect('/characters');
+		const figures = await Figure.find({ character: req.params.id });
+
+		// delete the charactr if there is no associated existing figure product
+		if(figures.length <= 0) {
+			await Character.findByIdAndRemove(req.params.id);
+			return res.status(204).redirect('/characters');
+		}
+
+		const character = await Character.findById(req.params.id);
+		res.render('character_delete', { 
+			title: 'Delete Character', 
+			figures,
+			character
+		});
 	} catch (err) {
 		next(err);
 	}
